requirements.txt
xnxx_api
av
httpx[http2]
colorama
pytest-cov